feat(api): Setup shooter profile controller
