+++ apps/api/package.json
@@ -10,9 +10,9 @@
 		"import": "./dist/index.js"
 	},
 	"scripts": {
-		"dev": "tsx watch src/index.ts",
+		"dev": "tsx watch src/server.ts",
 		"build": "tsc",
-		"start": "cross-env NODE_ENV=production node dist/index.js"
+		"start": "cross-env NODE_ENV=production node dist/server.js"
 	},
 	"keywords": [],
 	"author": "",
@@ -27,7 +27,8 @@
 		"better-auth-mikro-orm": "^0.4.3",
 		"dotenv": "^17.2.1",
 		"elysia": "^1.3.21",
-		"mikro-orm-soft-delete": "^1.0.0-alpha.1"
+		"mikro-orm-soft-delete": "^1.0.0-alpha.1",
+		"nodemailer": "^7.0.6"
 	},
 	"devDependencies": {
 		"@mikro-orm/cli": "^6.5.1",
