+++ apps/api/src/modules/shooterProfile/index.ts
@@ -1,0 +1,9 @@
+import { authPlugin, ormPlugin } from "@/plugins.js";
+import { Elysia } from "elysia";
+
+export const shooterProfileRoute = new Elysia({
+	prefix: "/shooter-profile",
+})
+	.use(ormPlugin)
+	.use(authPlugin)
+	.post("/", () => "eqweqw", { isAuth: true });
